Reorganized files into model/view folders, slight testing and modification of files
