+++ src/model/Main.java
@@ -1,4 +1,4 @@
-package main;
+package model;
 
 //TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
 // click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
