+++ src/model/dataEntry.java
@@ -1,4 +1,4 @@
-package main;
+package model;
 
 public class dataEntry {
     // This class manipulates the parsed data, making it easier to input into database
