+++ src/model/musicStore.java
@@ -1,4 +1,4 @@
-package main;
+package model;
 
 public class musicStore {
     // This class is where we implement all the data parsing within data added within our application
