+++ src/model/parserTests.java
@@ -1,4 +1,4 @@
-package main;
+package model;
 
 import org.junit.Test;
 
@@ -6,7 +6,7 @@ import static org.junit.Assert.*;
 
 public class parserTests {
     @Test
-    public void initialTest() {
+    public void testParserStatus() {
         parser test_parser = new parser();
         assertFalse(test_parser.parserStatus());
     }
